Fixing project installation on other devices

+++ database/migrations/2025_01_04_073639_add_foreign_key_to_orders_table.php
@@ -12,16 +12,19 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+            // Only add the 'user_id' column if it does not already exist
+            if (!Schema::hasColumn('orders', 'user_id')) {
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+            }
         });
     }
     
     public function down(): void
     {
         Schema::table('orders', function (Blueprint $table) {
+            // Drop foreign key constraint and column
             $table->dropForeign(['user_id']);
             $table->dropColumn('user_id');
         });
     }
-    
 };
